refactor(node)!: Move `Pnpm` into its own dedicated package
Prepare for adding further `Pnpm` specific classes.
